move config to lib and CONFIG to state
