+++ src/key_handlers.rs
@@ -12,16 +12,15 @@ use x11::xlib::{
     XUngrabServer, XWarpPointer, XWindowChanges, CWY,
 };
 
-use crate::config::CONFIG;
 use crate::enums::WM;
-use crate::{
+use crate::main_::{
     arrange, attach, attachstack, detach, detachstack, drawbar, focus,
     getrootptr, height, is_visible, nexttiled, pop, recttomon, resize,
     resizebarwin, restack, sendevent, setfullscreen, unfocus, updatebarpos,
-    width, xerror, xerrordummy, HANDLER, MOUSEMASK, SCRATCHTAG, TAGMASK, XNONE,
+    width, xerror, xerrordummy, HANDLER, MOUSEMASK, XNONE,
 };
-use rwm::State;
-use rwm::{Arg, Client, Monitor};
+use crate::{cfor, State};
+use crate::{Arg, Client, Monitor};
 
 pub(crate) fn togglebar(state: &mut State, _arg: *const Arg) {
     unsafe {
@@ -32,7 +31,7 @@ pub(crate) fn togglebar(state: &mut State, _arg: *const Arg) {
             [(*state.selmon).pertag.curtag as usize];
         updatebarpos(state, state.selmon);
         resizebarwin(state, state.selmon);
-        if CONFIG.showsystray {
+        if state.CONFIG.showsystray {
             let mut wc = XWindowChanges {
                 x: 0,
                 y: 0,
@@ -67,7 +66,8 @@ pub(crate) fn focusstack(state: &mut State, arg: *const Arg) {
         let mut i: *mut Client;
 
         if (*state.selmon).sel.is_null()
-            || ((*(*state.selmon).sel).isfullscreen && CONFIG.lock_fullscreen)
+            || ((*(*state.selmon).sel).isfullscreen
+                && state.CONFIG.lock_fullscreen)
         {
             return;
         }
@@ -167,7 +167,7 @@ pub(crate) fn zoom(state: &mut State, _arg: *const Arg) {
 pub(crate) fn view(state: &mut State, arg: *const Arg) {
     log::trace!("view");
     unsafe {
-        if (*arg).ui() & *TAGMASK
+        if (*arg).ui() & state.tagmask()
             == (*state.selmon).tagset[(*state.selmon).seltags as usize]
         {
             return;
@@ -176,9 +176,9 @@ pub(crate) fn view(state: &mut State, arg: *const Arg) {
 
         // Safety: we were gonna dereference it anyway
         let pertag = &mut (*state.selmon).pertag;
-        if ((*arg).ui() & *TAGMASK) != 0 {
+        if ((*arg).ui() & state.tagmask()) != 0 {
             (*state.selmon).tagset[(*state.selmon).seltags as usize] =
-                (*arg).ui() & *TAGMASK;
+                (*arg).ui() & state.tagmask();
             pertag.prevtag = pertag.curtag;
 
             if (*arg).ui() == !0 {
@@ -245,7 +245,7 @@ pub(crate) fn setlayout(state: &mut State, arg: *const Arg) {
         if arg.is_null()
             || (*arg).l().is_none()
             || !std::ptr::eq(
-                &CONFIG.layouts[(*arg).l().unwrap()],
+                &state.CONFIG.layouts[(*arg).l().unwrap()],
                 (*state.selmon).lt[(*state.selmon).sellt as usize],
             )
         {
@@ -258,7 +258,7 @@ pub(crate) fn setlayout(state: &mut State, arg: *const Arg) {
             (*state.selmon).pertag.ltidxs
                 [(*state.selmon).pertag.curtag as usize]
                 [(*state.selmon).sellt as usize] =
-                &CONFIG.layouts[(*arg).l().unwrap()];
+                &state.CONFIG.layouts[(*arg).l().unwrap()];
             (*state.selmon).lt[(*state.selmon).sellt as usize] =
                 (*state.selmon).pertag.ltidxs
                     [(*state.selmon).pertag.curtag as usize]
@@ -358,8 +358,9 @@ pub(crate) fn pushstack(state: &mut State, arg: *const Arg) {
 
 pub(crate) fn tag(state: &mut State, arg: *const Arg) {
     unsafe {
-        if !(*state.selmon).sel.is_null() && (*arg).ui() & *TAGMASK != 0 {
-            (*(*state.selmon).sel).tags = (*arg).ui() & *TAGMASK;
+        if !(*state.selmon).sel.is_null() && (*arg).ui() & state.tagmask() != 0
+        {
+            (*(*state.selmon).sel).tags = (*arg).ui() & state.tagmask();
             focus(state, null_mut());
             arrange(state, state.selmon);
         }
@@ -431,7 +432,7 @@ pub(crate) fn toggleview(state: &mut State, arg: *const Arg) {
     unsafe {
         let newtagset = (*state.selmon).tagset
             [(*state.selmon).seltags as usize]
-            ^ ((*arg).ui() & *TAGMASK);
+            ^ ((*arg).ui() & state.tagmask());
 
         if newtagset != 0 {
             (*state.selmon).tagset[(*state.selmon).seltags as usize] =
@@ -544,21 +545,25 @@ pub(crate) fn movemouse(state: &mut State, _arg: *const Arg) {
                     lasttime = ev.motion.time;
                     let mut nx = ocx + (ev.motion.x - x);
                     let mut ny = ocy + (ev.motion.y - y);
-                    if ((*state.selmon).wx - nx).abs() < CONFIG.snap as c_int {
+                    if ((*state.selmon).wx - nx).abs()
+                        < state.CONFIG.snap as c_int
+                    {
                         nx = (*state.selmon).wx;
                     } else if (((*state.selmon).wx + (*state.selmon).ww)
                         - (nx + width(c)))
                     .abs()
-                        < CONFIG.snap as c_int
+                        < state.CONFIG.snap as c_int
                     {
                         nx = (*state.selmon).wx + (*state.selmon).ww - width(c);
                     }
-                    if ((*state.selmon).wy - ny).abs() < CONFIG.snap as c_int {
+                    if ((*state.selmon).wy - ny).abs()
+                        < state.CONFIG.snap as c_int
+                    {
                         ny = (*state.selmon).wy;
                     } else if (((*state.selmon).wy + (*state.selmon).wh)
                         - (ny + height(c)))
                     .abs()
-                        < CONFIG.snap as c_int
+                        < state.CONFIG.snap as c_int
                     {
                         ny =
                             (*state.selmon).wy + (*state.selmon).wh - height(c);
@@ -568,8 +573,8 @@ pub(crate) fn movemouse(state: &mut State, _arg: *const Arg) {
                             .arrange
                             .0
                             .is_some()
-                        && ((nx - c.x).abs() > CONFIG.snap as c_int
-                            || (ny - c.y).abs() > CONFIG.snap as c_int)
+                        && ((nx - c.x).abs() > state.CONFIG.snap as c_int
+                            || (ny - c.y).abs() > state.CONFIG.snap as c_int)
                     {
                         togglefloating(state, null_mut());
                     }
@@ -670,8 +675,8 @@ pub(crate) fn resizemouse(state: &mut State, _arg: *const Arg) {
                             .arrange
                             .0
                             .is_some()
-                        && ((nw - c.w).abs() > CONFIG.snap as c_int
-                            || (nh - c.h).abs() > CONFIG.snap as c_int)
+                        && ((nw - c.w).abs() > state.CONFIG.snap as c_int
+                            || (nh - c.h).abs() > state.CONFIG.snap as c_int)
                     {
                         togglefloating(state, null_mut());
                     }
@@ -716,14 +721,14 @@ pub(crate) fn resizemouse(state: &mut State, _arg: *const Arg) {
 pub(crate) fn spawn(state: &mut State, arg: *const Arg) {
     unsafe {
         let mut argv = (*arg).v();
-        if argv == *CONFIG.dmenucmd {
+        if argv == *state.CONFIG.dmenucmd {
             log::trace!("spawn: dmenucmd on monitor {}", (*state.selmon).num);
             argv.push("-m".into());
             argv.push((*state.selmon).num.to_string());
         }
 
         (*state.selmon).tagset[(*state.selmon).seltags as usize] &=
-            !*SCRATCHTAG;
+            !state.scratchtag();
 
         let mut cmd = Command::new(argv[0].clone());
         let cmd = if argv.len() > 1 { cmd.args(&argv[1..]) } else { &mut cmd };
@@ -740,7 +745,8 @@ pub(crate) fn toggletag(state: &mut State, arg: *const Arg) {
         if (*state.selmon).sel.is_null() {
             return;
         }
-        let newtags = (*(*state.selmon).sel).tags ^ ((*arg).ui() & *TAGMASK);
+        let newtags =
+            (*(*state.selmon).sel).tags ^ ((*arg).ui() & state.tagmask());
         if newtags != 0 {
             (*(*state.selmon).sel).tags = newtags;
             focus(state, null_mut());
@@ -774,7 +780,7 @@ pub(crate) fn togglescratch(state: &mut State, arg: *const Arg) {
         c = (*state.selmon).clients;
         !c.is_null();
         c = (*c).next) {
-            found = ((*c).tags & *SCRATCHTAG) != 0;
+            found = ((*c).tags & state.scratchtag()) != 0;
             if found {
                 break;
             }
@@ -782,7 +788,7 @@ pub(crate) fn togglescratch(state: &mut State, arg: *const Arg) {
         if found {
             let newtagset = (*state.selmon).tagset
                 [(*state.selmon).seltags as usize]
-                ^ *SCRATCHTAG;
+                ^ state.scratchtag();
             if newtagset != 0 {
                 (*state.selmon).tagset[(*state.selmon).seltags as usize] =
                     newtagset;
